chore(eval): refresh all 27 server artifacts
Full /run-eval all pass at engine 0.9.0.dev1, skill v0.0.4-159-g8cc836b.
Regenerates wrappers, shapes, runners, session overviews, result.json and
doc/EVAL_REPORT.md for every server in servers.toml.

Verdicts: 21 pass, 6 partial. Partials are github, semgrep, notion, linear,
playwright (probe_inconclusive on tools with no observable success payload;
oauth roundtrip unsupported in the verifier) — firecrawl is now a full pass.

+++ eval/exa/exa.py
@@ -14,7 +14,7 @@ SERVER = 'exa'
 
 
 
-async def web_fetch_exa(caller: McpCaller, *, urls: list[str], maxCharacters: int | None = None) -> Any:
+async def web_fetch_exa(caller: McpCaller, *, urls: list[str], maxCharacters: float | None = None) -> Any:
     """Read a webpage's full content as clean markdown. Use after web_search_exa when highlights are insufficient or to read any URL.
 
     Best for: Extracting full content from known URLs. Batch multiple URLs in one call.
@@ -32,7 +32,7 @@ async def web_fetch_exa(caller: McpCaller, *, urls: list[str], maxCharacters: in
 web_fetch_exa.__schema__ = {'type': 'object', 'properties': {'urls': {'type': 'array', 'items': {'type': 'string'}, 'description': 'URLs to read. Batch multiple URLs in one call.'}, 'maxCharacters': {'type': 'number', 'minimum': 1, 'description': 'Maximum characters to extract per page (default: 3000)'}}, 'required': ['urls'], 'additionalProperties': False, '$schema': 'http://json-schema.org/draft-07/schema#'}
 
 
-async def web_search_exa(caller: McpCaller, *, query: str, numResults: int | None = None) -> Any:
+async def web_search_exa(caller: McpCaller, *, query: str, numResults: float | None = None) -> Any:
     """Search the web for any topic and get clean, ready-to-use content.
 
           Best for: Finding current information, news, facts, people, companies, or answering questions about any topic.
